FEATURE: HandleAnswers function
